// types.ts
export type UserType = {
  uid?: string;
  email?: string | null;
  name: string | null;
  number: string | null;
  address: string | null;
  role?: "user" | "manager" | "admin";
  isEmployee?: true | false;
  image?: any;
} | null;

export type UserDataType = {
  name: string;
  number: string;
  address: string;
  image?: any;
};

export type AuthContextType = {
  user: UserType;
  setUser: Function;
  login: (
    email: string,
    password: string
  ) => Promise<{ success: boolean; msg?: string }>;
  register: (
    email: string,
    password: string,
    number: string,
    address: string,
    name: string
  ) => Promise<{ success: boolean; msg?: string }>;
  updateUserData: (userId: string) => Promise<void>;
};